support upgrading GrapheneOS-specific settings
This is done to reduce the maintenance burden needed to deal with
upstream settings migrations and to also allow us to do any upgrades /
migrations with GrapheneOS-specific settings. It's only run if the
version changes, when new users are added---basically any other time a
normal setting upgrade would occur.

When some upgrade is needed (e.g., setting defaults reliably, or moving
from Secure to Global or vice versa), add the upgrade code in
onUpgradeGrapheneOsSettingsLocked, and update the
GRAPHENEOS_SETTINGS_VERSION field in SettingsProvider#UpdateController.

Upgrading of GrapheneOS-settings uses an independent version from the
normal settings version, thus upgrades are done independently (but done
after the normal settings upgrades). This means that we don't have to
worry as much about managing conflicts between GrapheneOS settings and
settings version upgrades from upstream.

For now, the GrapheneOS settings version is stored in Settings.Secure,
even though the normal settings version is stored as an XML attribute.