quick fixup for fingerprint unlock setting
This feature was developed during Android 10
and`UserHandle.getCallingUserId()` was used in the Android 10 version,
but I didn't change it to the Android 11 version which has a
`callingUserId` passed to the callback. From testing, there's no
functional difference between using `UserHandle.getCallingUserId()` and
using `callingUserId`, but this makes it more consistent with the
upstream face unlock enabled setting.