gmscompat: Optimize compatibility change checks
Querying the state of a compatibility change is relatively expensive
because it involves a layer of Compatibility.Callbacks indirection
followed by an array binary search for disabled change IDs. Considering
that some of our hooks are in potentially-hot paths, such as
File#lastModified() in file manager/archival apps, it's worth optimizing
the checks.

For performance, check the compatibility change states once at init time
(before Application#onCreate) and cache the states in volatile fields.
The fields need to be volatile in order to ensure thread safety during
early app init (especially if GmsCompat classes get loaded into Zygote
somehow), but that's a miniscule cost compared to calling
Compatibility#isChangeEnabled repeatedly.

Change-Id: Ic8a9afb388428efc0ead638bd6f41bc789192cbc