gmscompat: Check past signing certificates for forceQueryable
The force queryable check in AppsFilter calls isGmsApp with the
*current* APK signatures, which works for some Google apps. However,
Google Services Framework (com.google.android.gsf) has a v3 signature
that uses a new certificate, so the legacy MD5 certificate we check for
is classified as a past signing certificate instead. As a result,
forceQueryable is never set for GSF because it fails the isGmsApp check.

Account for both current and past signing certificates in the AppsFilter
path in order to fix the issue.

Change-Id: Id4f4002dcc4882e986707c9e02b970246ee62d77