gmscompat: add AppOps compatibility layer
Play services uses app ops to enforce system permissions for apps using
APIs through it. This can be supported for an untrusted proxy with the
proxy app op APIs. There's no equivalent to startOp for the proxy API
system so noteOp is used as an approximation where Play services will be
taking more blame than necessary, which is fine.