gmscompat: improve userId-related shims
Current approach is fragile: GMS always obtains UserHandle of the primary ("Owner") user and
tries to use it, which is a privileged operation if GMS is installed in the non-primary user profile.
It's worked around by unconditionally replacing UserHandle with the current one in all methods that
GMS calls, which will break each time new GMS version makes use of a new UserHandle-accepting method.

With the new approach, GMS gets access only to the current UserHandle which is made
to look like a UserHandle of the primary user.