disable monet by default
We need to add an opt-in toggle for this feature since we don't want it
enabled by default. We don't think people want the color scheme based on
their wallpaper by default. The resulting color scheme is accessible to
apps so it can be used to fingerprint a user based on their wallpaper,
including from a nested work profile.