CensoredNotification: Respect user switchability
Previous behaviro did not have checks on its pending intent whenever the
current user profile allows for user switchability and if multiple user
is enabled. Resolve this by removing ONE_SHOT pending intent flag to
allow tapping the notification action again the calls end and switch to
the user that sent the notification.

The check is based on exisiting UI on multiple user below quick tiles
and user settings:
https://cs.android.com/android/platform/superproject/+/master:frameworks/base/packages/SystemUI/src/com/android/systemui/statusbar/policy/UserSwitcherController.java;l=310;bpv=1;bpt=1

https://cs.android.com/android/platform/superproject/+/master:packages/apps/Settings/src/com/android/settings/users/UserDetailsSettings.java;drc=9f4eb48493739fb8975a236bce72865b8f82d5a7;bpv=1;bpt=1;l=330