gmscompat: improve telephony state listener shim
Allow listener to be registered if at least some of the events that it listens to are allowed
by permissions that GMS has.
Previously, if at least one event wasn't allowed, registration was skipped entirely.
Also, support registering listeners via the deprecated TelephonyManager#listen() method
(GMS still uses it in some cases), in addition to the modern registerTelephonyCallback().