[temporary] appops: add logging to catch construction of invalid ops
AppOpsManager.setMode() defining packageName as @Nullable seems to be a bug.
AppOpsService.setMode() that is wrapped by this method defines packageName as @NonNull, and code
that modifies AppOpsService state when handling this call assumes that packageName is not null.
This might be related to the bugs described in two previous commits.