gmscompat: prevent auto-install of "Play Services for AR" package
When this package is not installed, pretend that it's installed but disabled.
Then, if the user presses the "Enable" button, tell the Play Store that it was uninstalled.
This allows the user to install it manually when it's needed.
Client library for accessing "Play Services for AR" checks for its presence on startup and prompts
to install it if it's absent.