prevent Play services adding Wallet tile by itself
the Wallet tile is a system tile that checks if NFC is supported on the
device and if there's a registered default NFC payment provider using
the secure setting "nfc_payment_default_component". if both of these are
true, then the wallet tile will be reported as available to be added.

the tile can still be added manually and is still fully functional.

some users get confused or weirded out that this happens, and Google Pay
/ Wallet for NFC payments do not work on GrapheneOS at the moment so lets
just make them happy.

Signed-off-by: r3g_5z <june@girlboss.ceo>