[temporary] remove ScreenshotProxyService's dependency on CentralSurfaces
Workarounds issue that causes unexpected initialisation of
OverviewProxyService, when a screenshot is taken, resulting in broken
recents list and launcher crashes.

Note: When any screenshot action button (edit, share and scroll) from
ScreenshotView (which appears after screenshot is taken) is used on a
locked device, the keyguard needs to be unlocked to perform that action.
The dismissKeyguard() function is used, in these cases, to unlock the
keyguard before performing that action. This change intentionally breaks
dismissKeyguard() function, which causes the screenshot action buttons
to break too (only when keyguard is locked).

Test (manual):
1. Switch to secondary user profile (which is at rest).
2. Take screenshot and press any of the buttons in the popup.
3. Open recents list and make sure that it is working properly and the
   launcher isn't crashing.