add workaround for dynamic code logging system_server crash
PackageDynamicCodeLoading class is used for collecting info about dynamic code loading, which is
then logged to the SafetyNet EventLog. This log is currently not used on GrapheneOS.

The replaced statement gets executed, despite the comment saying that it should be always
unreachable, likely due to an upstream bug in entry collection logic.

See https://github.com/GrapheneOS/os-issue-tracker/issues/3017