require owner credential to check whether duress password is present
Previously, ACCESS_KEYGUARD_SECURE_STORAGE permission was enough to check that.

Presence of duress password is intended to be unknown unless the owner credential is provided. This
was enforced in the UI layer (Settings app) already, but could be bypassed through UI bugs, such
as the recently patched predictive back gesture bug:
https://github.com/GrapheneOS/platform_packages_apps_Settings/commit/e6ac6fa0f33f9b69cbcb4464fa46d93b548b0f2a